fix(netmgr)!: make NETCONN_CMD_PRI's payload the type it is stored in
netmgr.h described the payload as int; the drivers read *(int *)param and
assigned it to netmgr_conn_base_t.pri, a uint8_t, with no range check
anywhere. So pri = 300 became 44, and pri = -1 became 255 - the HIGHEST
priority, the exact opposite of what someone writing -1 means. Silent, and
nothing in the public header said which direction priority even ran.

The payload is now uint8_t on both the set and the get side, in all three
drivers. That is the whole fix: with the stored type as the payload type the
out-of-range values cannot be expressed, so there is nothing left to
validate and no truncation left to document. It also puts the command back
in line with its neighbours, every one of which names the type param points
at (NW_IP_S, netconn_wifi_info_t, netmgr_reconn_table_t) - int was the only
entry naming a transport type rather than the field's own.

This is the one change in this set the compiler cannot flag: param is void *,
so a caller still passing int * compiles. On every platform TuyaOpen supports
it also still works, because they are all little-endian and the first byte is
the one that matters for 0..255. On a big-endian port it would read zero and
drop that link to the lowest priority, silently. Documented in
release_notes 2.5 rather than left to be discovered.

Also documents the polarity where an application developer actually meets it.
"Higher wins" was stated in netconn_registry.h and netmgr_policy.h - the two
headers a board author and an extension author read - and nowhere in
netmgr.h. netmgr_conn_base_t.pri had no comment at all. Both now say higher
wins, with the built-in values (wired 2, wifi 1, cellular 0) so a reader can
check the claim without opening netconn_table.c, and default_pri now warns
that the tie-break beside it runs the other way: highest pri, then LOWEST
reg_index.

And corrects NETCONN_CMD_STATUS, whose payload comment said netmgr_type_e
while all three drivers write netmgr_status_e. Comment only; code unchanged.

BREAKING CHANGE: pass uint8_t * to netmgr_conn_set/get(NETCONN_CMD_PRI),
not int *.

+++ src/tuya_cloud_service/netmgr/src/conn/netconn_cellular.c
@@ -142,7 +142,7 @@ OPERATE_RET netconn_cellular_set(netmgr_conn_config_type_e cmd, void *param)
 
     switch (cmd) {
     case NETCONN_CMD_PRI: {
-        netmgr_cellular->base.pri = *(int *)param;
+        netmgr_cellular->base.pri = *(uint8_t *)param;
         /* Guarded like every other call to it in this file. Unreachable through
          * netmgr_conn_set(), which refuses before netmgr_init() has installed the
          * shim - but netconn_cellular_set() is on the global public include path, so
@@ -176,7 +176,7 @@ OPERATE_RET netconn_cellular_get(netmgr_conn_config_type_e cmd, void *param)
 
     switch (cmd) {
     case NETCONN_CMD_PRI: {
-        *(int *)param = netmgr_cellular->base.pri;
+        *(uint8_t *)param = netmgr_cellular->base.pri;
     } break;
     case NETCONN_CMD_STATUS: {
         *(netmgr_status_e *)param = netmgr_cellular->base.status;
